# swagger.yaml
openapi: 3.0.0
info:
  title: MEN REST API
  description: MongoDB ExpressJS NodeJS REST API
  version: 1.0.0
servers:
  - url: http://localhost:4000/api/
    description: Localhost development server 
components:
  securitySchemes:
    ApiKeyAuth:        # arbitrary name for the security scheme
      type: apiKey
      in: header       # can be "header", "query" or "cookie"
      name: auth-token

  schemas:
    TV-Show:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        seasons:
          type: number
        actors:
          type: string
    
    User:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        date:
          type: string

paths:

  /user/register:
    post:
      summary: Register a new User
      description: Register a new User in the database
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"

      responses:
        201:
          description: User created succesfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  _id:
                    type: string

  /user/login:
    post:
      summary: Login a User
      description: Login a User in order to get JWT token (auth-token)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      
      responses:
        200:
          description: User logged in succesfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  data:
                    type: string


  /tvshow:
    get:
      summary: Retrieves a list of tvshows
      description: Retrieves a list of tvshows as JSON objects.
      responses:
        200:
          description: A list of tvshows JSON objects in an array.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/tvshow"
    
    post:
      summary: Create a new tvshow
      description: Create a new tvshow
      security:
        - ApiKeyAuth: []      
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/tvshow"

      responses:
        201:
          description: tvshow created succesfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/tvshow"
  
  /tvshow/random:
    get:
      summary: Retrieves a random tvshow
      description: Retrieves a random tvshow as JSON objects.
      responses:
        200:
          description: A random tvshow JSON objects in an array.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/tvshow"

  /tvshow/{id}:
    get:
      summary: Specific tvshow
      description: Retrieves a specific tvshow based on it id.
      parameters:
        - in: path
          name: id
          required: true
          description: MongoDB id
          schema:
            type: string
      responses:
        200:
          description: A tvshow in the format of a JSON object.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/tvshow"
   
    put:
      summary: Updates a specific tvshow
      description: Updates a specific tvshow based on it id
      security:
        - ApiKeyAuth: []            
      parameters:
        - in: path
          name: id
          required: true
          description: MongoDB id
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/tvshow"

      responses:
        201:
          description: tvshow updated succesfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/tvshow"

    delete:
      summary: Deletes a specific tvshow
      description: Deletes a specific tvshow based on it id
      security:
        - ApiKeyAuth: []            
      parameters:
        - in: path
          name: id
          required: true
          description: MongoDB id
          schema:
            type: string      

      responses:
        201:
          description: tvshow deleted succesfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/tvshow"